+++ Vault/src/utopia/vault/util/ErrorHandling.scala
@@ -17,6 +17,7 @@ object ErrorHandling
 	var defaultPrinciple: ErrorHandlingPrinciple = Ignore
 	
 	private var _modelParsePrinciple: Option[ErrorHandlingPrinciple] = None
+	private var _insertClipPrinciple: Option[ErrorHandlingPrinciple] = None
 	
 	
 	// COMPUTED	---------------------
@@ -26,4 +27,12 @@ object ErrorHandling
 	  */
 	def modelParsePrinciple = _modelParsePrinciple.getOrElse(defaultPrinciple)
 	def modelParsePrinciple_=(principle: ErrorHandlingPrinciple) = _modelParsePrinciple = Some(principle)
+	
+	/**
+	  * @return Error handling principle used when properties are not included in an insert because they
+	  *         don't belong to a table, which may be result of a wrong property name definition
+	  */
+	//noinspection MutatorLikeMethodIsParameterless
+	def insertClipPrinciple = _insertClipPrinciple.getOrElse(defaultPrinciple)
+	def insertClipPrinciple_=(principle: ErrorHandlingPrinciple) = _insertClipPrinciple = Some(principle)
 }